+++ src/assets/scss/application.scss
@@ -25,6 +25,7 @@ $govuk-new-link-styles: true;
 @import "components/summary-list/summary-list";
 @import "components/warning-text/warning-text";
 @import "components/radios/radios";
+@import "components/table/table";
 // end Design System component imports
 
 // language toggle component import
